🚚build: 파일 분리
